refactor(chat): extract turn output finalizer
